requirements.txt
uvicorn
fastapi
psycopg2-binary 
